gap/backend/requirements.txt
Django
psycopg2
channels
Pillow
Redis
channels-redis
djangorestframework
django-filter
django-rest-swagger
django-cors-headers